+++ src/components/NavBar/NavBarProducts/NavBarProducts.jsx
@@ -12,7 +12,8 @@ import {
 
 import {
   fontWeightMedium,
-  fontFamily
+  fontFamily,
+  fontSizeSmall
 } from '../../style/fonts';
 
 const StyledNavBarProducts = styled.nav`
@@ -52,10 +53,12 @@ const NewLabel = styled.span`
   color: ${white};
   font-family: ${fontFamily};
   font-weight: ${fontWeightMedium};
+  font-size: ${fontSizeSmall};
   background: ${green};
-  border-radius: 4px;
+  border-radius: 12px;
   padding: 2px 8px;
   margin-left: 8px;
+  margin-top: 2px;
 `;
 
 // Notee: Custom SVGs to support multi-color product icons (instead of using icon component)