+++ src/components/style/colors.js
@@ -5,7 +5,7 @@ export const blueLight = '#6B81FF';
 export const blueLighter = '#ABB7FF';
 export const purple = '#9C2BFF';
 export const pink = '#D925AC';
-export const pinkLighter  = '#F0A8DE';
+export const pinkLighter = '#F0A8DE';
 export const teal = '#00C8CF';
 export const red = '#E0364F';
 export const redDark = '#9D2637';
@@ -29,17 +29,19 @@ export const grayDefault = '#77797a';
 export const boxShadow = '#ABB7FF';
 export const grayShadow = '0px 1px 4px rgba(0, 0, 0, 0.16)';
 
-export const facebook = '#4267B2';
-export const facebookDark = '#2B4373';
+export const facebook = '#1877F2';
+export const facebookDark = '#0C51AB';
 export const googleplus = '#dd4b39';
-export const instagram = '#BD3381';
-export const instagramDark = '#782152';
-export const linkedin = '#0077B5';
-export const linkedinDark = '#003C5C';
-export const pinterest = '#BD081C';
-export const pinterestDark = '#6C0510';
+export const instagram = '#F00276';
+export const instagramDark = '#A51D5F';
+export const linkedin = '#2867B2';
+export const linkedinDark = '#1A4476';
+export const pinterest = '#E60023';
+export const pinterestDark = '#BD081C';
+export const shopify = '#95BF47';
+export const shopifyDark = '#5E8E3E';
 export const twitter = '#1DA1F2';
-export const twitterDark = '#0A6FAE';
+export const twitterDark = '#1174B0';
 
 export const transparent = 'transparent';